Average of random number (Python code)
Generate random number between 0 to 100, for 100000 iteration. The mean is very very close of 50. which mean randint function follow the normal distribution as mean closes to 50. I ran this code twice is Python and the result is 49.9979167 and 49.9961673.

The result is some wayward when using Excel - RandBetween (1,100). - the average for 1048575 value comes to 50.43.